Crate gu_base

Modules

cli

files

Structs

App
Used to create a representation of a command line program and all possible command line arguments. Application settings are set using the "builder pattern" with the App::get_matches family of methods being the terminal methods that starts the runtime-parsing process. These methods then return information about the user supplied arguments (or lack there of).

Arg
The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgMatches
Used to get information about the arguments that where supplied to the program at runtime by the user. New instances of this struct are obtained by using the App::get_matches family of methods.

AutocompleteModule

ChainModule

GuApp

LogModule

SubCommand
The abstract representation of a command line subcommand.

Traits

Decorator

Module

ModuleChain

Functions

run_once